requirements.txt
# Core web scraping and HTTP
aiohttp==3.9.1
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
playwright==1.40.0
feedparser==6.0.10
requests==2.31.0

# API clients
praw==7.7.1
tweepy==4.14.0
requests-oauthlib==1.3.1

# Async and concurrency
aiofiles==23.2.1
asyncpg==0.29.0
motor==3.3.2
asyncio-throttle==1.0.2

# Data processing and validation
pandas==2.1.3
pydantic==2.5.0
pydantic-settings==2.0.3
python-dateutil==2.8.2

# Database and storage
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
psycopg2-binary==2.9.9

# Rate limiting and reliability
aiolimiter==1.1.0
backoff==2.2.1
tenacity==8.2.3

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
typer==0.9.0
rich==13.7.0

# Monitoring and logging
loguru==0.7.2
structlog==23.2.0
sentry-sdk==1.38.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
